refactor: make file name simple
