+++ pkg/adapter/controller/user.go
@@ -1,29 +1,29 @@
 package controller
 
 import (
+	"golang-clean-architecture/pkg/usecase/usecase"
 	"net/http"
 
 	"golang-clean-architecture/pkg/domain/model"
-	"golang-clean-architecture/pkg/usecase/interactor"
 )
 
 type userController struct {
-	userInteractor interactor.UserInteractor
+	userUsecase usecase.User
 }
 
-type UserController interface {
+type User interface {
 	GetUsers(c Context) error
 	CreateUser(c Context) error
 }
 
-func NewUserController(us interactor.UserInteractor) UserController {
+func NewUserController(us usecase.User) User {
 	return &userController{us}
 }
 
 func (uc *userController) GetUsers(c Context) error {
 	var u []*model.User
 
-	u, err := uc.userInteractor.Get(u)
+	u, err := uc.userUsecase.Get(u)
 	if err != nil {
 		return err
 	}
@@ -38,7 +38,7 @@ func (uc *userController) CreateUser(c Context) error {
 		return err
 	}
 
-	u, err := uc.userInteractor.Create(&params)
+	u, err := uc.userUsecase.Create(&params)
 	if err != nil {
 		return err
 	}
