+++ pkg/infrastructure/router/router.go
@@ -1,7 +1,7 @@
 package router
 
 import (
-	"golang-clean-architecture/pkg/interface/controller"
+	"golang-clean-architecture/pkg/adapter/controller"
 
 	"github.com/labstack/echo"
 	"github.com/labstack/echo/middleware"
