--- pkg/interface/presenter/user_presenter.go
@@ -1,19 +1,0 @@
-package presenter
-
-import (
-	"golang-clean-architecture/pkg/domain/model"
-	"golang-clean-architecture/pkg/usecase/presenter"
-)
-
-type userPresenter struct{}
-
-func NewUserPresenter() presenter.UserPresenter {
-	return &userPresenter{}
-}
-
-func (up *userPresenter) ResponseUsers(us []*model.User) []*model.User {
-	for _, u := range us {
-		u.Name = "Mr." + u.Name
-	}
-	return us
-}
