--- pkg/usecase/presenter/user_presenter.go
@@ -1,7 +1,0 @@
-package presenter
-
-import "golang-clean-architecture/pkg/domain/model"
-
-type UserPresenter interface {
-	ResponseUsers(u []*model.User) []*model.User
-}
